+++ integration_tests/contracts/contracts/TestERC20Utility.sol
@@ -1,4 +1,4 @@
-pragma solidity 0.8.10;
+pragma solidity 0.8.21;
 import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 
 // An utility erc20 contract that has a fancy method