+++ integration_tests/contracts/hardhat.config.ts
@@ -4,11 +4,12 @@ module.exports = {
   solidity: {
     compilers: [
       {
-        version: "0.8.10",
+        version: "0.8.21",
         settings: {
           optimizer: {
             enabled: true
-          }
+          },
+          evmVersion: "shanghai"
         }
       },
       {