+++ integration_tests/test_ibc_timeout.py
@@ -17,7 +17,7 @@ def ibc(request, tmp_path_factory):
     "prepare-network"
     name = "ibc_timeout"
     path = tmp_path_factory.mktemp(name)
-    yield from prepare_network(path, name)
+    yield from prepare_network(path, name, grantee="signer3")
 
 
 def test_ibc(ibc):